more API 13 fixes

+++ VFXEditor/DirectX/Renderers/ModelRenderer.cs
@@ -305,7 +305,7 @@ namespace VfxEditor.DirectX.Renderers {
             Resize( size );
 
             var pos = ImGui.GetCursorScreenPos();
-            ImGui.ImageButton( Output, size, new Vec2( 0, 0 ), new Vec2( 1, 1 ), 0 );
+            ImGui.ImageButton( (ulong)Output, size, new Vec2( 0, 0 ), new Vec2( 1, 1 ), 0 );
 
             var topLeft = ImGui.GetItemRectMin();
             var bottomRight = ImGui.GetItemRectMax();
